// src/types/GameAverages.ts
export interface GameAverages {
  gp: number;
  gs: number;
  timePlayed: string;
  fgm: number;
  fga: number;
  fgPercent: number | null;
  tpm: number;
  tpa: number;
  tpPercent: number | null;
  ftm: number;
  fta: number;
  ftPercent: number | null;
  oreb: number;
  dreb: number;
  reb: number;
  ast: number;
  stl: number;
  blk: number;
  tov: number;
  pf: number;
  pts: number;
  plusMinus: number;
  rn: number;
}